# internal/server/openapi.yaml
openapi: 3.0.3
info:
  title: TeamVault API
  version: 0.9.1
  description: |
    Zero-Knowledge Passwortmanager REST-API.
    Der Server speichert nur Ciphertext + Envelopes; Ver-/Entschlüsselung
    erfolgt ausschließlich in WebUI, Extension oder CLI (cryptocore).

    API-Key-Scopes (mind. einer): `read` (GET-Allowlist), `vault` (Secret-/Vault-Mutationen),
    `admin` (`/api/admin/*`, plus User-Rollen). Cookie-Sessions ohne Scope-Limit.

    Instanz-Backup: `GET /api/admin/backup`, `POST /api/admin/backup/restore` (platform_admin, Ciphertext only).
servers:
  - url: http://127.0.0.1:8080
    description: Local default
...
security:
  - sessionCookie: []
  - bearerAPIKey: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tv_session
    bearerAPIKey:
      type: http
      scheme: bearer
      bearerFormat: tvk_
      description: |
        API-Key aus Admin-UI (einmalig angezeigt). Gebunden an User+Tenant.
        Scopes: read | vault | admin (mind. einer).
...
paths:
...
  /api/admin/backup/restore:
    post:
      tags: [admin]
      summary: Replace vault store from snapshot; confirm=RESTORE
      responses:
        '200':
          description: Restore applied
        '400':
          description: confirm missing or snapshot invalid
...